requirements.txt
environs==9.5.0
python-telegram-bot==13.14
google-cloud-dialogflow==2.16.1
google-cloud-storage==2.5.0
gcloud==0.18.3